Add setting to automatically detect indentation when loading a file
This is necessary because the autocmd must be set before custom filetype
plugins are loaded.  Otherwise, a file can be opened which should
receive a filetype that has C-style comments, but detectindent will not
know about the filetype yet, so it will not be marked as having C-style
comments, which could result in incorrect indentation being detected.