Make the `g:detectindent_preferred_expandtab` value be a boolean
… rather than assuming the variable is true because it exists, no matter its value

In other words, treat a `g:detectindent_preferred_expandtab` 0 value to signify a desired default of `noexpandtab`.

This allows the user to `let g:detectindent_preferred_expandtab = 0`, signifying that they definitely want `noexpandtab` if the detection is unsure. It makes the choice of the suggested “1” value in the documentation less arbitrary.