Extract function SetLocalIndentWidth(num_spaces)
Makes the code shorter and simpler.

This does sometimes set tabstop or softtabstop when it’s unnecessary to do so (because expandtab is the wrong value), but there is no harm in doing that. And in fact it is helpful, because it reduces the work the user has to do if they manually change the value of expandtab – the option will already be at the correct value.