Add pointer to look at docs in usage comments
I feel like it might be a bad practice to duplicate the docs in the function comments, so I don’t want to redescribe the function options. But maybe there was a good reason for putting the comments there. So I compromise by instructing the developer to look at the docs.

At least now the usage comments are clearly labeled incomplete, so a developer won’t think that the two options mentioned are the only ones.