Temporize that the original only “seems to be” abandoned
It has a bunch of open issues with no comments, but the creator didn’t officially abandon it, and someone still filed an issue there a few months ago. So it still has some activity, though the plugin itself isn’t being updated.

The original’s creator hasn’t explicitly said it, but a comment in a closed issue makes me think that they don’t want the lenient, advanced mixed tabs/spaces resolution, preferring to make life hard for those who edit mixed files to encourage fixing it. If this is so, then even if the original is updated, I should describe that difference in my fork, so as to attract others who want smarter detection.