Flesh out the README with description and more
The description is based on the description of this plugin in the Vim.org plugin repository, at http://www.vim.org/scripts/script.php?script_id=1171.

I wrote this months ago, but hadn’t committed it. I also wrote a section “About the detection algorithm”, but I’m not committing it yet because it’s not finished.