Add script tests/run-fixtures-test
The reason I might want to run only this test is that this is the only test that is easily applicable to older versions of DetectIndent, which have fewer features. I want to use the fixtures test to compare accuracy between the old and new algorithms.

I did some refactoring so that this script did not duplicate much code with run-all-tests.