Improve augroup definition code
I added the missing `!` reset, and improved the if check by changing `exists` to `get`, so that the code is not run if the option exists but is set to 0.

I also improved the augroup name. I can use capitals because `:help autocmd-groups` says the names are case-sensitive. The augroup name now matches the capitalization in the README. (I still need to either document or remove this autocommand definition in the plugin code.)

In the README, I expanded the `autocmd!` command name, so it is clearer what the command does, and consistent with the next line.