Split body of DetectIndent() into two sections
Now 'expandtab' and the indent width are set separately. This is easier to read and better-factored. It also fixes a bug where the indent width of the file was not detected if tabs were detected in the file.

This refactoring has also revealed a possibly-redundant option and room for further refactoring, and I have noted those possibilities for change in comments.