Extract function RestrictIndentWidthWithinOptionRange
As I asked for in a TODO comment

I am choosing to pass the indent width in a parameter for now, rather than reading it from a shim of shiftwidth(). I am doing this because I have a vague suspicion that the current indent width might not always equal the shiftwidth(), if I make one of the changes requested in the issue tracker, which I think involves having some of the indent-width options set different from each other.