Make example test output match output for users
The test output in the README included output for a test in progress, `config-options-are-respected.vader`, which is not committed in the repository. Now the example test output is the same as the output seen by users who clone this repository and run the tests. This should make it easier for users to confirm that any modifications they make didn’t break a test.